Wasn't a bug, just needed to move a camera. Terrain finished

+++ GeometryGenerator.cpp
@@ -22,23 +22,23 @@ void GeometryGenerator::CreateGrid(float width, float depth, UINT m, UINT n, Mes
 		for (UINT j = 0; j < n; ++j)
 		{
 			float x = -halfWidth + j * dx;
-			meshData.Vertices[static_cast<INT>((static_cast<INT>(i) * static_cast<INT>(n) + static_cast<INT>(j)))].Position = DirectX::XMFLOAT3(x, 0.0f, z);
+			meshData.Vertices[(long long)i * n + (long long)j].Position = DirectX::XMFLOAT3(x, 0.0f, z);
 		}
 	}
 
-	meshData.Indices.resize(triangleCount * static_cast<INT>(3));
+	meshData.Indices.resize((long long)triangleCount * 3);
 	UINT k = 0;
 	for (UINT i = 0; i < m - 1; ++i)
 	{
 		for (UINT j = 0; j < n - 1; ++j)
 		{
 			meshData.Indices[k] = i * n + j;
-			meshData.Indices[k + 1] = i * n + j + 1;
-			meshData.Indices[k + 2] = (i + 1) * n + j;
+			meshData.Indices[(long long)k + 1] = i * n + j + 1;
+			meshData.Indices[(long long)k + 2] = (i + 1) * n + j;
 
-			meshData.Indices[k + 3] = (i + 1) * n + j;
-			meshData.Indices[k + 4] = i * n + j + 1;
-			meshData.Indices[k + 5] = (i + 1) * n + j + 1;
+			meshData.Indices[(long long)k + 3] = (i + 1) * n + j;
+			meshData.Indices[(long long)k + 4] = i * n + j + 1;
+			meshData.Indices[(long long)k + 5] = (i + 1) * n + j + 1;
 			k += 6; // next quad
 		}
 	}
